# backend/docs/openapi.v1.yaml
openapi: 3.0.3
info:
  title: bimo API
  version: 1.0.0
servers:
  - url: http://localhost:8001
paths:
...
  /v1/admin/ingest/invoice:
    post:
      summary: Admin ingest invoice
      parameters:
        - in: header
          name: X-Admin-Token
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                invoice:
                  type: object
                  additionalProperties: true
      responses:
        "200": { description: OK }
...